Backport timing configuration changes to 2.0.x

+++ tck/src/main/java/org/eclipse/microprofile/fault/tolerance/tck/timeout/clientserver/UninterruptableTimeoutClient.java
@@ -54,7 +54,7 @@ public class UninterruptableTimeoutClient {
      * @param waitMs the time to wait
      */
     @Timeout(value = 500, unit = ChronoUnit.MILLIS)
-    public void serviceTimeout(int waitMs) {
+    public void serviceTimeout(long waitMs) {
         long waitNs = Duration.ofMillis(waitMs).toNanos();
         long startTime = System.nanoTime();
